server/requirements_server.txt
Flask
pymongo
python-dotenv # For managing environment variables
bcrypt # For password hashing
Flask-Login # For session management